Forgot to actually save the session in createSession
Also on a related note I added a doc comment to this, and some more text
to the comment for `hostname`. I'm wondering if that's actually
important at all. The only situation it could protect against is one
where someone has another user's session ID, and wants to make a request
against a different sub-site. And that's if I even do the sub-sites
thing. It just seems pretty meaningless, extra hassle for no benefit.